requirements.txt
absl-py
altair
annotated-types
anyio
astunparse
attrs
beautifulsoup4
blinker
bs4
cachetools
certifi
charset-normalizer
click
colorama
contourpy
cycler
defusedxml
distro
duckduckgo_search
et_xmlfile
flatbuffers
fonttools
gast
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
google-pasta
googleapis-common-protos
grpcio
grpcio-status
h11
h5py
httpcore
httplib2
httpx
idna
Jinja2
jiter
joblib
jsonschema
jsonschema-specifications
keras
kiwisolver
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
namex
narwhals
numpy
openai
openpyxl
opt_einsum
optree
packaging
pandas
pillow
primp
proto-plus
protobuf
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic_core
pydeck
Pygments
pyparsing
python-dateutil
python-dotenv
python-pptx
pytz
referencing
requests
rich
rpds-py
rsa
scikit-learn
scipy
seaborn
six
smmap
sniffio
soupsieve
streamlit
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
toml
tornado
tqdm
typing_extensions
tzdata
uritemplate
urllib3
watchdog
websockets
Werkzeug
wrapt
XlsxWriter
youtube-transcript-api
